Refactor: removed duplicate code
